moved exceptions to errors
